Issue #11163: Enforced file size on MethodName check inputs
