+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/sizes/MethodLengthCheckTest.java
@@ -84,6 +84,20 @@ public class MethodLengthCheckTest extends AbstractModuleTestSupport {
     }
 
     @Test
+    public void testCountEmptyIsFalseTwo() throws Exception {
+        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
+        verifyWithInlineConfigParser(
+                getPath("InputMethodLengthCountEmptyIsFalseTwo.java"), expected);
+    }
+
+    @Test
+    public void testCountEmptyIsFalseThree() throws Exception {
+        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
+        verifyWithInlineConfigParser(
+                getPath("InputMethodLengthCountEmptyIsFalseThree.java"), expected);
+    }
+
+    @Test
     public void testWithComments() throws Exception {
         final String[] expected = {
             "35:5: " + getCheckMessage(MSG_KEY, 8, 7, "visit"),
